Add signature in SAML request for service provider authentication at IdPs
An IdP can authenticate a service provider when a signature is included in SAMLRequest.
However, current nginx-lua-saml-service-provider does not include it and IdP cannot authenticate the signature in SAMLRequest.
This patch adds the signature to SAMLRequest.
A private key file for the signature is specified by "config.request.sp_private_key" in init.lua.
Note that this patch supports only RSA-SHA256 as a signature algorithm.